alterada a base da aplicacao para blueprints v1
