+++ config.py
@@ -3,6 +3,9 @@ import os
 # Define o caminho raiz do projeto
 basedir = os.path.abspath(os.path.dirname(__file__))
 
-# Define o caminho completo para o arquivo do banco de dados
-DATABASE_PATH = os.path.join(basedir, 'aplicacao.db')
-SQLALCHEMY_DATABASE_URI = 'sqlite:///' + DATABASE_PATH
+class Config:
+    SECRET_KEY = os.environ.get('SECRET_KEY') or 'uma-chave-secreta-muito-dificil-de-adivinhar'
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
+        'sqlite:///' + os.path.join(basedir, 'aplicacao.db')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    UPLOAD_FOLDER = os.path.join(basedir, 'mentoria', 'static', 'uploads')
