+++ mentoria/admin/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+admin = Blueprint('admin', __name__)
+from . import routes
