+++ mentoria/main/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+main = Blueprint('main', __name__)
+from . import routes
