+++ mentoria/templates/admin/admin_dashboard.html
@@ -32,16 +32,16 @@
 <!-- Secção de Acesso Rápido -->
 <h3 class="mb-3">Acesso Rápido</h3>
 <div class="list-group">
-    <a href="{{ url_for('admin_lista_inscricoes') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
+    <a href="{{ url_for('admin.admin_lista_inscricoes') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
         <div><i class="bi bi-people-fill me-2"></i><strong>Gerir Inscrições</strong></div><i class="bi bi-chevron-right"></i>
     </a>
-    <a href="{{ url_for('admin_lista_turmas') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
+    <a href="{{ url_for('admin.admin_lista_turmas') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
         <div><i class="bi bi-collection-fill me-2"></i><strong>Gerir Turmas</strong></div><i class="bi bi-chevron-right"></i>
     </a>
-    <a href="{{ url_for('admin_lista_modulos') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
+    <a href="{{ url_for('admin.admin_lista_modulos') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
         <div><i class="bi bi-collection-play-fill me-2"></i><strong>Gerir Módulos e Conteúdos</strong></div><i class="bi bi-chevron-right"></i>
     </a>
-    <a href="{{ url_for('admin_lista_monitores') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
+    <a href="{{ url_for('admin.admin_lista_monitores') }}" class="list-group-item list-group-item-action d-flex justify-content-between align-items-center">
         <div><i class="bi bi-person-video me-2"></i><strong>Gerir Monitores</strong></div><i class="bi bi-chevron-right"></i>
     </a>
 </div>
