+++ mentoria/templates/admin/detalhe.html
@@ -39,12 +39,12 @@
     <div class="card-footer text-end">
         <!-- NOVO BOTÃO PARA ADMINS -->
         {% if current_user.is_authenticated and current_user.is_admin %}
-        <a href="{{ url_for('definir_senha_mentorado', id=registro.id) }}" class="btn btn-warning">
+        <a href="{{ url_for('admin.definir_senha_mentorado', id=registro.id) }}" class="btn btn-warning">
             <i class="bi bi-key-fill"></i> Definir Senha
         </a>
         {% endif %}
         
-        <a href="{{ url_for('index') }}" class="btn btn-primary"><i class="bi bi-arrow-left"></i> Voltar para a Lista</a>
+        <a href="{{ url_for('main.index') }}" class="btn btn-primary"><i class="bi bi-arrow-left"></i> Voltar para a Lista</a>
     </div>
     
 </div>
