+++ mentoria/templates/admin/detalhes_modulo.html
@@ -3,7 +3,7 @@
 {% block content %}
 <div class="d-flex justify-content-between align-items-center">
     <div><h1 class="mb-0">{{ modulo.titulo }}</h1><p class="lead text-muted">{{ modulo.descricao }}</p></div>
-    <a href="{{ url_for('admin_lista_modulos') }}" class="btn btn-secondary"><i class="bi bi-arrow-left"></i> Voltar para Módulos</a>
+    <a href="{{ url_for('admin.admin_lista_modulos') }}" class="btn btn-secondary"><i class="bi bi-arrow-left"></i> Voltar para Módulos</a>
 </div>
 <hr>
 <div class="row g-5">
@@ -14,7 +14,7 @@
             <div class="list-group-item">
                 <div class="d-flex w-100 justify-content-between">
                     <div><h5 class="mb-1"><span class="badge bg-secondary me-2">{{ conteudo.ordem }}</span>{{ conteudo.titulo }}</h5><small class="badge bg-info text-dark">{{ conteudo.tipo }}</small></div>
-                    <div><a href="{{ url_for('admin_editar_conteudo', id=conteudo.id) }}" class="btn btn-sm btn-outline-secondary" title="Editar"><i class="bi bi-pencil"></i></a><form action="{{ url_for('admin_deletar_conteudo', id=conteudo.id) }}" method="POST" class="d-inline" onsubmit="return confirm('Tem a certeza?');"><button type="submit" class="btn btn-sm btn-outline-danger" title="Remover"><i class="bi bi-trash"></i></button></form></div>
+                    <div><a href="{{ url_for('admin.admin_editar_conteudo', id=conteudo.id) }}" class="btn btn-sm btn-outline-secondary" title="Editar"><i class="bi bi-pencil"></i></a><form action="{{ url_for('admin.admin_deletar_conteudo', id=conteudo.id) }}" method="POST" class="d-inline" onsubmit="return confirm('Tem a certeza?');"><button type="submit" class="btn btn-sm btn-outline-danger" title="Remover"><i class="bi bi-trash"></i></button></form></div>
                 </div>
                 <p class="mb-1 mt-2">{{ conteudo.descricao }}</p>
                 <small><a href="{{ conteudo.url_conteudo }}" target="_blank">{{ conteudo.url_conteudo }}</a></small>
@@ -28,7 +28,7 @@
         <h3>Adicionar Novo Conteúdo</h3>
         <div class="card">
             <div class="card-body">
-                <form action="{{ url_for('admin_novo_conteudo', modulo_id=modulo.id) }}" method="POST" enctype="multipart/form-data">
+                <form action="{{ url_for('admin.admin_novo_conteudo', modulo_id=modulo.id) }}" method="POST" enctype="multipart/form-data">
                     <div class="mb-3"><label class="form-label">Título</label><input type="text" name="titulo" class="form-control" required></div>
                     <div class="mb-3"><label class="form-label">URL (Link do Drive, YouTube, etc.)</label><input type="url" name="url_conteudo" class="form-control" placeholder="https://..." required></div>
                     <div class="row"><div class="col-sm-7"><label class="form-label">Tipo</label><select name="tipo" class="form-select"><option>Vídeo</option><option>PDF</option><option>Slides</option><option>Link Externo</option></select></div><div class="col-sm-5"><label class="form-label">Ordem</label><input type="number" name="ordem" class="form-control" value="0"></div></div>
