+++ mentoria/templates/admin/form_conteudo.html
@@ -53,7 +53,7 @@
             </div>
             <hr>
             <div class="d-flex justify-content-end">
-                <a href="{{ url_for('admin_detalhes_modulo', id=conteudo.modulo_id) }}" class="btn btn-secondary me-2">Cancelar</a>
+                <a href="{{ url_for('admin.admin_detalhes_modulo', id=conteudo.modulo_id) }}" class="btn btn-secondary me-2">Cancelar</a>
                 <button type="submit" class="btn btn-primary">Salvar Alterações</button>
             </div>
         </form>
