+++ mentoria/templates/admin/form_definir_senha.html
@@ -16,7 +16,7 @@
             </div>
             <hr>
             <div class="d-flex justify-content-end">
-                <a href="{{ url_for('admin_detalhe_inscricao', id=mentorado.id) }}" class="btn btn-secondary me-2">Cancelar</a>
+                <a href="{{ url_for('admin.admin_detalhe_inscricao', id=mentorado.id) }}" class="btn btn-secondary me-2">Cancelar</a>
                 <button type="submit" class="btn btn-primary">Salvar Nova Senha</button>
             </div>
         </form>
