+++ mentoria/templates/admin/form_encontro.html
@@ -52,7 +52,7 @@
             <hr>
 
             <div class="d-flex justify-content-end">
-                <a href="{{ url_for('admin_detalhes_turma', id=turma.id) }}" class="btn btn-secondary me-2">Cancelar</a>
+                <a href="{{ url_for('admin.admin_detalhes_turma', id=turma.id) }}" class="btn btn-secondary me-2">Cancelar</a>
                 <button type="submit" class="btn btn-primary">
                     <i class="bi bi-check-circle"></i> Salvar Encontro
                 </button>
