+++ mentoria/templates/admin/form_modulo.html
@@ -12,7 +12,7 @@
                 <div class="col-md-9 mb-3"><label for="thumbnail_url" class="form-label">URL da Imagem de Capa (Opcional)</label><input type="url" class="form-control" id="thumbnail_url" name="thumbnail_url" value="{{ modulo.thumbnail_url if modulo else '' }}" placeholder="[https://exemplo.com/imagem.jpg](https://exemplo.com/imagem.jpg)"></div>
             </div>
             <hr>
-            <div class="d-flex justify-content-end"><a href="{{ url_for('admin_lista_modulos') }}" class="btn btn-secondary me-2">Cancelar</a><button type="submit" class="btn btn-primary">Salvar Módulo</button></div>
+            <div class="d-flex justify-content-end"><a href="{{ url_for('admin.admin_lista_modulos') }}" class="btn btn-secondary me-2">Cancelar</a><button type="submit" class="btn btn-primary">Salvar Módulo</button></div>
         </form>
     </div>
 </div>
