+++ mentoria/templates/admin/form_monitor.html
@@ -29,7 +29,7 @@
             </div>
             <hr>
             <div class="d-flex justify-content-end">
-                <a href="{{ url_for('admin_lista_monitores') }}" class="btn btn-secondary me-2">Cancelar</a>
+                <a href="{{ url_for('admin.admin_lista_monitores') }}" class="btn btn-secondary me-2">Cancelar</a>
                 <button type="submit" class="btn btn-primary"><i class="bi bi-check-circle"></i> Salvar</button>
             </div>
         </form>
