+++ mentoria/templates/admin/lista_modulos.html
@@ -3,7 +3,7 @@
 {% block content %}
 <div class="d-flex justify-content-between align-items-center mb-3">
     <h1 class="mb-0">Gerir Módulos</h1>
-    <a href="{{ url_for('admin_novo_modulo') }}" class="btn btn-primary"><i class="bi bi-plus-circle"></i> Criar Novo Módulo</a>
+    <a href="{{ url_for('admin.admin_novo_modulo') }}" class="btn btn-primary"><i class="bi bi-plus-circle"></i> Criar Novo Módulo</a>
 </div>
 <div class="card shadow-sm">
     <div class="card-body">
@@ -17,9 +17,9 @@
                         <td><strong>{{ modulo.titulo }}</strong></td>
                         <td>{{ modulo.descricao|truncate(80) }}</td>
                         <td class="text-end">
-                            <a href="{{ url_for('admin_detalhes_modulo', id=modulo.id) }}" class="btn btn-sm btn-outline-primary" title="Gerir Conteúdos"><i class="bi bi-plus-square"></i> Conteúdos</a>
-                            <a href="{{ url_for('admin_editar_modulo', id=modulo.id) }}" class="btn btn-sm btn-outline-secondary" title="Editar Módulo"><i class="bi bi-pencil-square"></i> Editar</a>
-                            <form action="{{ url_for('admin_deletar_modulo', id=modulo.id) }}" method="POST" class="d-inline" onsubmit="return confirm('Tem a certeza? Apagar um módulo também removerá todos os seus conteúdos.');"><button type="submit" class="btn btn-sm btn-outline-danger" title="Remover Módulo"><i class="bi bi-trash"></i></button></form>
+                            <a href="{{ url_for('admin.admin_detalhes_modulo', id=modulo.id) }}" class="btn btn-sm btn-outline-primary" title="Gerir Conteúdos"><i class="bi bi-plus-square"></i> Conteúdos</a>
+                            <a href="{{ url_for('admin.admin_editar_modulo', id=modulo.id) }}" class="btn btn-sm btn-outline-secondary" title="Editar Módulo"><i class="bi bi-pencil-square"></i> Editar</a>
+                            <form action="{{ url_for('admin.admin_deletar_modulo', id=modulo.id) }}" method="POST" class="d-inline" onsubmit="return confirm('Tem a certeza? Apagar um módulo também removerá todos os seus conteúdos.');"><button type="submit" class="btn btn-sm btn-outline-danger" title="Remover Módulo"><i class="bi bi-trash"></i></button></form>
                         </td>
                     </tr>
                     {% else %}
