+++ mentoria/templates/admin/lista_monitores.html
@@ -3,7 +3,7 @@
 {% block content %}
 <div class="d-flex justify-content-between align-items-center mb-3">
     <h1 class="mb-0">Gerir Monitores</h1>
-    <a href="{{ url_for('admin_novo_monitor') }}" class="btn btn-primary">
+    <a href="{{ url_for('admin.admin_novo_monitor') }}" class="btn btn-primary">
         <i class="bi bi-plus-circle"></i> Registar Novo Monitor
     </a>
 </div>
@@ -26,10 +26,10 @@
                         <td>{{ monitor.email }}</td>
                         <td><span class="badge bg-secondary">{{ monitor.perfil }}</span></td>
                         <td class="text-end">
-                            <a href="{{ url_for('admin_editar_monitor', id=monitor.id) }}" class="btn btn-sm btn-outline-secondary">
+                            <a href="{{ url_for('admin.admin_editar_monitor', id=monitor.id) }}" class="btn btn-sm btn-outline-secondary">
                                 <i class="bi bi-pencil-square"></i> Editar
                             </a>
-                            <form action="{{ url_for('admin_deletar_monitor', id=monitor.id) }}" method="POST" class="d-inline" onsubmit="return confirm('Tem a certeza que deseja remover este monitor?');">
+                            <form action="{{ url_for('admin.admin_deletar_monitor', id=monitor.id) }}" method="POST" class="d-inline" onsubmit="return confirm('Tem a certeza que deseja remover este monitor?');">
                                 <button type="submit" class="btn btn-sm btn-outline-danger">
                                     <i class="bi bi-trash"></i> Remover
                                 </button>
