+++ mentoria/templates/layout.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link rel="icon" type="image/x-icon" sizes="32x32" href="../static/img/icone_desenvolvedor.ico">
+    <link rel="icon" type="image/x-icon" sizes="32x32" href="../mentoria/static/img/icone_desenvolvedor.ico">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
     
@@ -82,7 +82,7 @@
     
     <nav class="navbar navbar-expand-lg navbar-dark shadow-sm">
         <div class="container">
-            <a class="navbar-brand" href="{{ url_for('index') }}">
+            <a class="navbar-brand" href="{{ url_for('main.index') }}">
                 <i class="bi bi-bar-chart-line-fill"></i>
                 Mentoria
             </a>
@@ -94,15 +94,15 @@
                     {% if current_user.is_authenticated %}
                         {% if current_user.is_admin %}
                         <li class="nav-item">
-                            <a class="nav-link text-warning" href="{{ url_for('admin_dashboard') }}">
+                            <a class="nav-link text-warning" href="{{ url_for('admin.admin_dashboard') }}">
                                 <i class="bi bi-gear-fill"></i> Painel Admin
                             </a>
                         </li>
                         {% endif %}
-                        <li class="nav-item"><a class="nav-link" href="{{ url_for('area_membros') }}"><i class="bi bi-person-circle"></i> Área de Membros</a></li>
-                        <li class="nav-item"><a class="nav-link" href="{{ url_for('logout') }}"><i class="bi bi-box-arrow-right"></i> Sair</a></li>
+                        <li class="nav-item"><a class="nav-link" href="{{ url_for('main.area_membros') }}"><i class="bi bi-person-circle"></i> Área de Membros</a></li>
+                        <li class="nav-item"><a class="nav-link" href="{{ url_for('main.logout') }}"><i class="bi bi-box-arrow-right"></i> Sair</a></li>
                     {% else %}
-                        <li class="nav-item"><a class="nav-link" href="{{ url_for('login') }}"><i class="bi bi-box-arrow-in-right"></i> Login</a></li>
+                        <li class="nav-item"><a class="nav-link" href="{{ url_for('main.login') }}"><i class="bi bi-box-arrow-in-right"></i> Login</a></li>
                     {% endif %}
                 </ul>
             </div>
