+++ mentoria/templates/login.html
@@ -8,7 +8,7 @@
         <div class="card shadow-sm">
             <div class="card-body p-4">
                 <h3 class="card-title text-center mb-4">Acessar Área de Membros</h3>
-                <form method="POST" action="{{ url_for('login') }}">
+                <form method="POST" action="{{ url_for('main.login') }}">
                     <div class="mb-3">
                         <label for="email" class="form-label">Email</label>
                         <input type="email" class="form-control" id="email" name="email" required>
