--- wsgi.py
@@ -1,5 +1,0 @@
-# wsgi.py
-from api.index import app
-
-if __name__ == "__main__":
-    app.run()
